backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
django-smart-selects>=1.6.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
